+++ crates/vela_engine/src/tests/source_reload/dir_function_abi.rs
@@ -258,9 +258,9 @@ fn runtime_stages_dir_native_parameter_rejection_until_safe_point() {
     let kind = dir_native_rejection_kind(
         "runtime_stage_dir_native_parameter",
         NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-            .param("amount", TypeHint::Int),
+            .param("amount", TypeHint::i64()),
         NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-            .param("amount", TypeHint::Float),
+            .param("amount", TypeHint::f64()),
         "reload.function.parameter_abi_changed",
     );
 
@@ -276,10 +276,10 @@ fn runtime_stages_dir_native_parameter_rejection_until_safe_point() {
     assert_eq!(function, "game::native::grant_bonus");
     assert_eq!(old.len(), 1);
     assert_eq!(old[0].name, "amount");
-    assert_eq!(old[0].type_hint.as_deref(), Some("int"));
+    assert_eq!(old[0].type_hint.as_deref(), Some("i64"));
     assert_eq!(new.len(), 1);
     assert_eq!(new[0].name, "amount");
-    assert_eq!(new[0].type_hint.as_deref(), Some("float"));
+    assert_eq!(new[0].type_hint.as_deref(), Some("f64"));
     assert!(source_span.is_none());
 }
 
@@ -288,9 +288,9 @@ fn runtime_stages_dir_native_return_rejection_until_safe_point() {
     let kind = dir_native_rejection_kind(
         "runtime_stage_dir_native_return",
         NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-            .returns(TypeHint::Int),
+            .returns(TypeHint::i64()),
         NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-            .returns(TypeHint::Float),
+            .returns(TypeHint::f64()),
         "reload.function.return_abi_changed",
     );
 
@@ -304,8 +304,8 @@ fn runtime_stages_dir_native_return_rejection_until_safe_point() {
         panic!("expected changed native function return ABI");
     };
     assert_eq!(function, "game::native::grant_bonus");
-    assert_eq!(old.as_deref(), Some("int"));
-    assert_eq!(new.as_deref(), Some("float"));
+    assert_eq!(old.as_deref(), Some("i64"));
+    assert_eq!(new.as_deref(), Some("f64"));
     assert!(source_span.is_none());
 }
 
@@ -362,7 +362,7 @@ pub fn grant() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-                .returns(TypeHint::Int)
+                .returns(TypeHint::i64())
                 .effects(EffectSet::host_read()),
             |_| Ok(OwnedValue::Scalar(vela_common::ScalarValue::I64(5))),
         )
@@ -375,7 +375,7 @@ pub fn grant() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus_v2", NativeFunctionId::new(22))
-                .returns(TypeHint::Int)
+                .returns(TypeHint::i64())
                 .effects(EffectSet::host_read()),
             |_| Ok(OwnedValue::Scalar(vela_common::ScalarValue::I64(5))),
         )
@@ -600,9 +600,9 @@ fn runtime_stages_dir_method_parameter_rejection_until_safe_point() {
     let kind = dir_method_rejection_kind(
         "runtime_stage_dir_method_parameter",
         MethodDesc::new(HostMethodId::new(9), "grant_exp")
-            .param(MethodParamDesc::new("amount").type_hint("int")),
+            .param(MethodParamDesc::new("amount").type_hint("i64")),
         MethodDesc::new(HostMethodId::new(9), "grant_exp")
-            .param(MethodParamDesc::new("amount").type_hint("float")),
+            .param(MethodParamDesc::new("amount").type_hint("f64")),
         "reload.method.parameter_abi_changed",
     );
 
@@ -620,10 +620,10 @@ fn runtime_stages_dir_method_parameter_rejection_until_safe_point() {
     assert_eq!(method, "grant_exp");
     assert_eq!(old.len(), 1);
     assert_eq!(old[0].name, "amount");
-    assert_eq!(old[0].type_hint.as_deref(), Some("int"));
+    assert_eq!(old[0].type_hint.as_deref(), Some("i64"));
     assert_eq!(new.len(), 1);
     assert_eq!(new[0].name, "amount");
-    assert_eq!(new[0].type_hint.as_deref(), Some("float"));
+    assert_eq!(new[0].type_hint.as_deref(), Some("f64"));
     assert!(source_span.is_none());
 }
 
@@ -631,7 +631,7 @@ fn runtime_stages_dir_method_parameter_rejection_until_safe_point() {
 fn runtime_stages_dir_method_return_rejection_until_safe_point() {
     let kind = dir_method_rejection_kind(
         "runtime_stage_dir_method_return",
-        MethodDesc::new(HostMethodId::new(9), "grant_exp").return_type("int"),
+        MethodDesc::new(HostMethodId::new(9), "grant_exp").return_type("i64"),
         MethodDesc::new(HostMethodId::new(9), "grant_exp").return_type("null"),
         "reload.method.return_abi_changed",
     );
@@ -648,7 +648,7 @@ fn runtime_stages_dir_method_return_rejection_until_safe_point() {
     };
     assert_eq!(type_name, "Player");
     assert_eq!(method, "grant_exp");
-    assert_eq!(old.as_deref(), Some("int"));
+    assert_eq!(old.as_deref(), Some("i64"));
     assert_eq!(new.as_deref(), Some("null"));
     assert!(source_span.is_none());
 }
